requirements.txt.txt
python-telegram-bot==20.7
gspread
google-auth
google-api-python-client
deep-translator